+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.2] - 16/04/2021
+
+* Added `backgroundColor` optional parameter with a default value of `Colors.white`
+* Added `iconColor` optional parameter with a default value of `Colors.black`
+
 ## [0.0.1] - 16/04/2021
 
 * Initial package release